+++ Rules/IPv4/Outbound/Development/Github.ps1
@@ -67,14 +67,17 @@ $GithubRoot = "%SystemDrive%\Users\User\AppData\Local\GitHubDesktop\app-2.2.3"
 # Test if installation exists on system
 if ((Test-Installation "Git" ([ref] $GitRoot) @Logs) -or $ForceLoad)
 {
+	# Administrators are needed for scheduled task for git auto update
+	$CurlUsers = Get-SDDL -Group "Administrators", "Users" @Logs
 	$Program = "$GitRoot\mingw64\bin\curl.exe"
 	Test-File $Program @Logs
+
 	New-NetFirewallRule -Platform $Platform `
 		-DisplayName "Git - curl" -Service Any -Program $Program `
 		-PolicyStore $PolicyStore -Enabled True -Action Allow -Group $Group -Profile $Profile -InterfaceType $Interface `
 		-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 443 `
-		-LocalUser $UsersGroupSDDL `
-		-Description "curl download tool" @Logs | Format-Output @Logs
+		-LocalUser $CurlUsers `
+		-Description "curl download tool, also used by Git for Windows Updater" @Logs | Format-Output @Logs
 
 	# TODO: unsure if it's 443 or 80
 	$Program = "$GitRoot\mingw64\bin\git.exe"
@@ -93,7 +96,7 @@ if ((Test-Installation "Git" ([ref] $GitRoot) @Logs) -or $ForceLoad)
 		-PolicyStore $PolicyStore -Enabled True -Action Allow -Group $Group -Profile $Profile -InterfaceType $Interface `
 		-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 443 `
 		-LocalUser $UsersGroupSDDL `
-		-Description "git HTTPS acces (https cloning)" @Logs | Format-Output @Logs
+		-Description "git HTTPS access (https cloning)" @Logs | Format-Output @Logs
 
 	$Program = "$GitRoot\usr\bin\ssh.exe"
 	Test-File $Program @Logs
@@ -102,7 +105,7 @@ if ((Test-Installation "Git" ([ref] $GitRoot) @Logs) -or $ForceLoad)
 		-PolicyStore $PolicyStore -Enabled False -Action Allow -Group $Group -Profile $Profile -InterfaceType $Interface `
 		-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 22 `
 		-LocalUser $UsersGroupSDDL `
-		-Description "git SSH acces" @Logs | Format-Output @Logs
+		-Description "git SSH access" @Logs | Format-Output @Logs
 }
 
 #