+++ Rules/IPv4/Outbound/Software/Intel.ps1
@@ -76,4 +76,25 @@ if ((Test-Installation "XTU" ([ref] $IntelXTURoot) @Logs) -or $ForceLoad)
 		-Description "Extreme Tuning utility check for updates" @Logs | Format-Output @Logs
 }
 
+$Program = "%ProgramFiles(x86)%\Intel\Telemetry 2.0\lrio.exe"
+Test-File $Program @Logs
+
+New-NetFirewallRule -Platform $Platform `
+	-DisplayName "Intel telemetry" -Service Any -Program $Program `
+	-PolicyStore $PolicyStore -Enabled True -Action Block -Group $Group -Profile $Profile -InterfaceType $Interface `
+	-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 80, 443 `
+	-LocalUser $NT_AUTHORITY_System `
+	-Description "Uploader for the Intel(R) Product Improvement Program." @Logs | Format-Output @Logs
+
+# TODO: port and protocol unknown for Intel PTT EK Recertification
+$Program = "%ProgramFiles%\Intel\Intel(R) Management Engine Components\iCLS\IntelPTTEKRecertification.exe"
+Test-File $Program @Logs
+
+New-NetFirewallRule -Platform $Platform `
+	-DisplayName "Intel PTT EK Recertification" -Service Any -Program $Program `
+	-PolicyStore $PolicyStore -Enabled True -Action Block -Group $Group -Profile $Profile -InterfaceType $Interface `
+	-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort Any `
+	-LocalUser $NT_AUTHORITY_System `
+	-Description "" @Logs | Format-Output @Logs
+
 Update-Logs